# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  authentication.AuthResponse:
    properties:
      accessToken:
        type: string
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      exp:
        type: integer
      extend: {}
      first_name:
        type: string
      id:
        type: integer
      last_login:
        type: string
      last_name:
        type: string
      phone:
        type: string
      role_id:
        type: integer
      role_name:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  authentication.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  authentication.ForgotPasswordRequest:
    properties:
      email:
        example: john@example.com
        type: string
    required:
    - email
    type: object
  authentication.LoginRequest:
    description: Login request payload
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  authentication.RegisterRequest:
    description: Registration request payload
    properties:
      email:
        description: '@Description User''s email address'
        example: john@example.com
        type: string
      first_name:
        description: '@Description User''s first name'
        example: John
        type: string
      last_name:
        description: '@Description User''s last name'
        example: Doe
        type: string
      password:
        description: '@Description Password for the account (minimum 8 characters)'
        example: password123
        minLength: 8
        type: string
      phone:
        description: '@Description User''s phone number'
        example: "+1234567890"
        type: string
      username:
        description: '@Description Username for the account'
        example: johndoe
        type: string
    required:
    - email
    - password
    type: object
  authentication.ResetPasswordRequest:
    properties:
      email:
        example: john@example.com
        type: string
      new_password:
        example: newpassword123
        minLength: 6
        type: string
      token:
        type: string
    required:
    - email
    - new_password
    - token
    type: object
  authentication.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact:
    email: info@base.al
    name: Base Team
    url: https://base.al
  description: This is the API documentation for Base Framework
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://base.al/terms
  title: Base Framework API
  version: 2.0.0
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request to reset password
      parameters:
      - description: Forgot Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authentication.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authentication.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Forgot Password
      tags:
      - Core/Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: Login Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authentication.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authentication.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Login
      tags:
      - Core/Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authentication.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - Core/Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register user
      parameters:
      - description: Register Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authentication.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/authentication.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Register
      tags:
      - Core/Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using token
      parameters:
      - description: Reset Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/authentication.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authentication.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authentication.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Reset Password
      tags:
      - Core/Auth
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key for authentication
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: Enter your token with the prefix "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"